+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ContactController;
+use App\Http\Controllers\ArticleController;
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\NewsController;
 use App\Http\Controllers\PageController;
@@ -15,6 +16,12 @@ Route::get('/', HomeController::class)->name('home');
 
 Route::get('/actualites', [NewsController::class, 'index'])->name('news.index');
 Route::get('/actualites/{slug}', [NewsController::class, 'show'])->name('news.show');
+
+if (\App\Support\Modules::enabled('articles')) {
+    Route::get('/article.php', [ArticleController::class, 'legacy'])->name('articles.legacy');
+    Route::get('/' . config('maracuja.articles.public_path', 'articles'), [ArticleController::class, 'index'])->name('articles.index');
+    Route::get('/' . config('maracuja.articles.public_path', 'articles') . '/{slug}', [ArticleController::class, 'show'])->name('articles.show');
+}
 
 Route::get('/contact', [ContactController::class, 'create'])->name('contact');
 Route::post('/contact', [ContactController::class, 'store'])->name('contact.store');